move transforms out of compose file (#1623)
move transforms out of compose file. this is backwards compatible.
