+++ monai/transforms/transform.py
@@ -1,0 +1,206 @@
+# Copyright 2020 - 2021 MONAI Consortium
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#     http://www.apache.org/licenses/LICENSE-2.0
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""
+A collection of generic interfaces for MONAI transforms.
+"""
+
+from abc import ABC, abstractmethod
+from typing import Any, Hashable, Optional, Tuple
+
+import numpy as np
+
+from monai.config import KeysCollection
+from monai.utils import MAX_SEED, ensure_tuple
+
+__all__ = ["Randomizable", "Transform", "MapTransform"]
+
+
+class Randomizable(ABC):
+    """
+    An interface for handling random state locally, currently based on a class variable `R`,
+    which is an instance of `np.random.RandomState`.
+    This is mainly for randomized data augmentation transforms. For example::
+
+        class RandShiftIntensity(Randomizable):
+            def randomize():
+                self._offset = self.R.uniform(low=0, high=100)
+            def __call__(self, img):
+                self.randomize()
+                return img + self._offset
+
+        transform = RandShiftIntensity()
+        transform.set_random_state(seed=0)
+
+    """
+
+    R: np.random.RandomState = np.random.RandomState()
+
+    def set_random_state(
+        self, seed: Optional[int] = None, state: Optional[np.random.RandomState] = None
+    ) -> "Randomizable":
+        """
+        Set the random state locally, to control the randomness, the derived
+        classes should use :py:attr:`self.R` instead of `np.random` to introduce random
+        factors.
+
+        Args:
+            seed: set the random state with an integer seed.
+            state: set the random state with a `np.random.RandomState` object.
+
+        Raises:
+            TypeError: When ``state`` is not an ``Optional[np.random.RandomState]``.
+
+        Returns:
+            a Randomizable instance.
+
+        """
+        if seed is not None:
+            _seed = id(seed) if not isinstance(seed, (int, np.integer)) else seed
+            _seed = _seed % MAX_SEED
+            self.R = np.random.RandomState(_seed)
+            return self
+
+        if state is not None:
+            if not isinstance(state, np.random.RandomState):
+                raise TypeError(f"state must be None or a np.random.RandomState but is {type(state).__name__}.")
+            self.R = state
+            return self
+
+        self.R = np.random.RandomState()
+        return self
+
+    @abstractmethod
+    def randomize(self, data: Any) -> None:
+        """
+        Within this method, :py:attr:`self.R` should be used, instead of `np.random`, to introduce random factors.
+
+        all :py:attr:`self.R` calls happen here so that we have a better chance to
+        identify errors of sync the random state.
+
+        This method can generate the random factors based on properties of the input data.
+
+        Raises:
+            NotImplementedError: When the subclass does not override this method.
+
+        """
+        raise NotImplementedError(f"Subclass {self.__class__.__name__} must implement this method.")
+
+
+class Transform(ABC):
+    """
+    An abstract class of a ``Transform``.
+    A transform is callable that processes ``data``.
+
+    It could be stateful and may modify ``data`` in place,
+    the implementation should be aware of:
+
+        #. thread safety when mutating its own states.
+           When used from a multi-process context, transform's instance variables are read-only.
+        #. ``data`` content unused by this transform may still be used in the
+           subsequent transforms in a composed transform.
+        #. storing too much information in ``data`` may not scale.
+
+    See Also
+
+        :py:class:`monai.transforms.Compose`
+    """
+
+    @abstractmethod
+    def __call__(self, data: Any):
+        """
+        ``data`` is an element which often comes from an iteration over an
+        iterable, such as :py:class:`torch.utils.data.Dataset`. This method should
+        return an updated version of ``data``.
+        To simplify the input validations, most of the transforms assume that
+
+        - ``data`` is a Numpy ndarray, PyTorch Tensor or string
+        - the data shape can be:
+
+          #. string data without shape, `LoadImage` transform expects file paths
+          #. most of the pre-processing transforms expect: ``(num_channels, spatial_dim_1[, spatial_dim_2, ...])``,
+             except that `AddChannel` expects (spatial_dim_1[, spatial_dim_2, ...]) and
+             `AsChannelFirst` expects (spatial_dim_1[, spatial_dim_2, ...], num_channels)
+          #. most of the post-processing transforms expect
+             ``(batch_size, num_channels, spatial_dim_1[, spatial_dim_2, ...])``
+
+        - the channel dimension is not omitted even if number of channels is one
+
+        This method can optionally take additional arguments to help execute transformation operation.
+
+        Raises:
+            NotImplementedError: When the subclass does not override this method.
+
+        """
+        raise NotImplementedError(f"Subclass {self.__class__.__name__} must implement this method.")
+
+
+class MapTransform(Transform):
+    """
+    A subclass of :py:class:`monai.transforms.Transform` with an assumption
+    that the ``data`` input of ``self.__call__`` is a MutableMapping such as ``dict``.
+
+    The ``keys`` parameter will be used to get and set the actual data
+    item to transform.  That is, the callable of this transform should
+    follow the pattern:
+
+        .. code-block:: python
+
+            def __call__(self, data):
+                for key in self.keys:
+                    if key in data:
+                        # update output data with some_transform_function(data[key]).
+                    else:
+                        # do nothing or some exceptions handling.
+                return data
+
+    Raises:
+        ValueError: When ``keys`` is an empty iterable.
+        TypeError: When ``keys`` type is not in ``Union[Hashable, Iterable[Hashable]]``.
+
+    """
+
+    def __init__(self, keys: KeysCollection) -> None:
+        self.keys: Tuple[Hashable, ...] = ensure_tuple(keys)
+        if not self.keys:
+            raise ValueError("keys must be non empty.")
+        for key in self.keys:
+            if not isinstance(key, Hashable):
+                raise TypeError(f"keys must be one of (Hashable, Iterable[Hashable]) but is {type(keys).__name__}.")
+
+    @abstractmethod
+    def __call__(self, data):
+        """
+        ``data`` often comes from an iteration over an iterable,
+        such as :py:class:`torch.utils.data.Dataset`.
+
+        To simplify the input validations, this method assumes:
+
+        - ``data`` is a Python dictionary
+        - ``data[key]`` is a Numpy ndarray, PyTorch Tensor or string, where ``key`` is an element
+          of ``self.keys``, the data shape can be:
+
+          #. string data without shape, `LoadImaged` transform expects file paths
+          #. most of the pre-processing transforms expect: ``(num_channels, spatial_dim_1[, spatial_dim_2, ...])``,
+             except that `AddChanneld` expects (spatial_dim_1[, spatial_dim_2, ...]) and
+             `AsChannelFirstd` expects (spatial_dim_1[, spatial_dim_2, ...], num_channels)
+          #. most of the post-processing transforms expect
+             ``(batch_size, num_channels, spatial_dim_1[, spatial_dim_2, ...])``
+
+        - the channel dimension is not omitted even if number of channels is one
+
+        Raises:
+            NotImplementedError: When the subclass does not override this method.
+
+        returns:
+            An updated dictionary version of ``data`` by applying the transform.
+
+        """
+        raise NotImplementedError(f"Subclass {self.__class__.__name__} must implement this method.")
