+++ tests/test_compose.py
@@ -167,6 +167,9 @@ class TestCompose(unittest.TestCase):
         # test len
         self.assertEqual(len(t1), 8)
 
+    def test_backwards_compatible_imports(self):
+        from monai.transforms.compose import MapTransform, Randomizable, Transform  # noqa: F401
+
 
 if __name__ == "__main__":
     unittest.main()
